Struct axhal::arch::TaskContext

pub struct TaskContext {
    pub kstack_top: VirtAddr,
    pub rsp: u64,
    pub fs_base: usize,
    pub ext_state: ExtendedState,
}

Saved hardware states of a task.

The context usually includes:

Callee-saved registers

Stack pointer register

Thread pointer register (for thread-local storage, currently unsupported)

FP/SIMD registers

On context switch, current task saves its context from CPU to memory, and the next task restores its context from memory to CPU.

On x86_64, callee-saved registers are saved to the kernel stack by the PUSH instruction. So that rsp is the RSP after callee-saved registers are pushed, and kstack_top is the top of the kernel stack (RSP before any push).

Fields
kstack_top: VirtAddr

The kernel stack top of the task.

rsp: u64

RSP after all callee-saved registers are pushed.

fs_base: usize

Thread Local Storage (TLS).

ext_state: ExtendedState
Available on crate feature fp_simd only.

Extended states, i.e., FP/SIMD states.

Implementations

impl TaskContext

pub const fn new() -> Self

Creates a new default context for a new task.

pub fn init(&mut self, entry: usize, kstack_top: VirtAddr, tls_area: VirtAddr)

Initializes the context for a new task, with the given entry point and kernel stack.

pub fn switch_to(&mut self, next_ctx: &Self)

Switches to another task.

It first saves the current task’s context from CPU to this place, and then restores the next task’s context from next_ctx to CPU.