axhal::arch
Struct UspaceContext

pub struct UspaceContext(/* private fields */);

Available on crate feature uspace only.

Context to enter user space.

Implementations

impl UspaceContext

pub const fn empty() -> Self

Creates an empty context with all registers set to zero.

pub fn new(entry: usize, ustack_top: VirtAddr, arg0: usize) -> Self

Creates a new context with the given entry point, user stack pointer, and the argument.

pub const fn from(tf: &TrapFrame) -> Self

Creates a new context from the given TrapFrame.

It copies almost all registers except CS and SS which need to be set to the user segment selectors.

pub unsafe fn enter_uspace(&self, kstack_top: VirtAddr) -> !

Enters user space.

It restores the user registers and jumps to the user entry point (saved in rip). When an exception or syscall occurs, the kernel stack pointer is switched to kstack_top.

Safety

This function is unsafe because it changes processor mode and the stack.