axhal

Module asm

Wrapper functions for assembly instructions.

Functions

disable_irqs

Makes the current CPU to ignore interrupts.

enable_irqs

Allows the current CPU to respond to interrupts.

flush_tlb

Flushes the TLB.

halt

Halt the current CPU.

irqs_enabled

Returns whether the current CPU is allowed to respond to interrupts.

read_kernel_page_table

Reads the current page table root register for kernel space (CR3).

read_thread_pointer

Reads the thread pointer of the current CPU (FS_BASE).

read_user_page_table

Reads the current page table root register for user space (CR3).

wait_for_irqs

Relaxes the current CPU and waits for interrupts.

write_kernel_page_table⚠

Writes the register to update the current page table root for kernel space (CR3).

write_thread_pointer⚠

Writes the thread pointer of the current CPU (FS_BASE).

write_user_page_table⚠

Writes the register to update the current page table root for user space (CR3).