Module axhal::arch

Architecture-specific types and operations.

Structs

ExtendedState

Extended state of a task, such as FP/SIMD states.

FxsaveArea

A 512-byte memory region for the FXSAVE/FXRSTOR instruction to save and restore the x87 FPU, MMX, XMM, and MXCSR registers.

GdtStruct

A wrapper of the Global Descriptor Table (GDT) with maximum 16 entries.

IdtStruct

A wrapper of the Interrupt Descriptor Table (IDT).

TaskContext

Saved hardware states of a task.

TaskStateSegment

In 64-bit mode the TSS holds information that is not directly related to the task-switch mechanism, but is used for finding kernel level stack if interrupts arrive while in kernel mode.

TrapFrame

Saved registers when a trap (interrupt or exception) occurs.

Functions

disable_irqs

Makes the current CPU to ignore interrupts.

enable_irqs

Allows the current CPU to respond to interrupts.

flush_tlb

Flushes the TLB.

halt

Halt the current CPU.

irqs_enabled

Returns whether the current CPU is allowed to respond to interrupts.

read_page_table_root

Reads the register that stores the current page table root.

read_thread_pointer

Reads the thread pointer of the current CPU.

wait_for_irqs

Relaxes the current CPU and waits for interrupts.

write_page_table_root⚠

Writes the register to update the current page table root.

write_thread_pointer⚠

Writes the thread pointer of the current CPU.